Extracted the pymcbdsc_root_dir method
from "__init__.py" to "utils.py".
