+++ tests/pymcbdsc/__init__.py
@@ -8,16 +8,12 @@ import shutil
 import random
 from sys import version_info
 from logging import getLogger
+# os_name2root_dir: os.name で取得できる OS の名前と、各 OS のデフォルトとなる pymbdsc_root_dir のデフォルト値のペア。
+# os_name2test_root_dir: os.name で取得できる OS の名前と、各 OS でのテストケース実行時に利用するテスト用ディレクトリパスのペア。
+from .test_utils import os_name2root_dir, os_name2test_root_dir
 
 
 logger = getLogger(__name__)
-
-# os.name で取得できる OS の名前と、各 OS のデフォルトとなる pymbdsc_root_dir のデフォルト値のペア。
-os_name2root_dir = {"posix": "/var/lib/pymcbdsc",
-                    "nt": "c:\\pymcbdsc"}
-# os.name で取得できる OS の名前と、各 OS でのテストケース実行時に利用するテスト用ディレクトリパスのペア。
-os_name2test_root_dir = {"posix": "/tmp/test/pymcbdsc",
-                         "nt": "c:\\test\\pymcbdsc"}
 
 
 def stop_patcher(patcher, patcher_name=None):
@@ -57,22 +53,6 @@ def create_empty_files(dir: str, files: List[str]) -> None:
     for file in files:
         with open(os.path.join(dir, file), "w") as f:
             f.write('')
-
-
-class TestCommon(unittest.TestCase):
-    """ pymcbdsc モジュールに定義されている関数をテストするテストクラス。 """
-
-    def _test_pymcbdsc_root_dir(self, os_name: str, exp: str) -> None:
-        p = mock.patch('pymcbdsc.os_name', os_name)
-        p.start()
-
-        act = pymcbdsc.pymcbdsc_root_dir()
-        self.assertEqual(act, exp)
-        p.stop()
-
-    def test_pymcbdsc_root_dir(self) -> None:
-        for (os_name, exp_root_dir) in os_name2root_dir.items():
-            self._test_pymcbdsc_root_dir(os_name, exp_root_dir)
 
 
 class TestMcbdscDownloader(unittest.TestCase):
